Adding preventClickToHighlight JS Event, HTML Parsing Regex
I added a very short mousedown handler for JavaScript, which prevents
double/triple-click to highlight on the target element. This may be
useful for elements used as buttons or toggles. It does not prevent
highlighting entirely, so the user can still click and drag to highlight
the given text. For me, this was desireable as my button was a row of a
table, but you may want a different solution.

I added a scary-looking regex for parsing HTML. Yes, I know, I get the
joke. I did seriously do this, however. It's real.. I'm pretty sure it
works. Now I have the option of using regex101.com or whatever to do
quick edits to HTML when otherwise I'd have to find a different tool.

I also moved the appendChildren file into the JavaScript folder where it
belongs. Nothing changed within it.