validateType.js: Validates and throws a pre-baked TypeError
A function which can validate against both primitives and
classes/prototype functions. If the type of the first parameter matches
the one specified by the second parameter, the first parameter is
returned.

Otherwise, a pre-baked TypeError is thrown, which specifies the type
expected vs the type found. You can also pass a string for the optional
third parameter, to also include a variable name in the error message.

This is mostly a convenience for obligate-vanilla projects which would
be doing dynamic type checking anyways. If you want a more complete
solution to type validation in your project, I recommend using tools
like Flow or transpiled languages like TypeScript.