Add buffer, callbackCompose and EventQueue, other minor changes
Added buffer, a function which accepts a function and returns a wrapped
function which has the same behavior, but will only execute at a
specified frequency. Additional calls made between executions will be
buffered, and executed in order later on. This is similar to
throttle.js, except with an arbitrary-size argument buffer.

callbackCompose is a function for chaining together callback functions
which accept other callbacks. This is one way to avoid callback hell.

EventQueue is just a basic class which lets you keep a list of functions
that can be executed in sequence at any time. Possibly useful for added
"subcriptions" or "listeners" to a callback after that callback has
already been passed to its caller.

intToLabel.js and throttle.js received no functional changes, mainly
getting slightly better comments.

preventClickToHighlight.js was changed to a function which accepts an
element and attaches the previously given event listener to it. It also
now returns the element for convenience.