Fixing icon link on index, minor css changes to output box
Made it so that the icon link on index also maintains the clean link
(see last commit).

Made the min-height of the output container 2em so that it matches the
height of the box when it contains a single line. The means the box
won't jump heights when typing your first character.