Cleaning up appendChildren/MakeElement, making callbackCompose use callback-last, adding pausableCB
pausableCB represents an object which normally just immediately calls any functions added to it. If it is in a paused state, the functions are queued up. When the pauser is unpaused, the functions will be called in the order passed in.

I intend to use this in JSCrawler to make it possible to pause it.