Adding insertionDistance, linearLex, lexicalCompare, range. some other changes made
A bunch of changes since I waited too long to commit this stuff.

linearLex and linearLex2 are differentiated mainly by the fact that the second supports specifying a sequence of tokens to follow a given one, rather than just being able to specify options fo rthe very next token.

lexicalCompare is totally unrelated to either of those (blame English being overloaded). It has to do with comparing arrays of numbers.