linearLex Open Close
Made a new version of linearLex for supporting parenthetical operators. Grammars can now specify new `open` and `close` properties on token types. Whenever a matching tag is met, if the token type has an `open` property, the value of the property (which may be any primitive, or a lazy-evaluated function, which is passed the regex match object for the token) is pushed to a stack. When a closing parenthetical operator is met, its `close` property value is checked against the top of the stack. If there's a mismatch (or the stack is empty), an error is thrown explaining that there is no matching opening operator for the given closing operator. If by the end of the parameter string the stack is not empty, a different error is thrown explaining that some opening operator(s) were not closed.

Another new feature in this version is the `DEFAULT_NEXT` grammar property, which allows users to specify the default legal token types when a given token doesn't specify its `next` property.

Also, all non-token grammar properties are now capitalized by convention, and properties on the grammar object are only treated as token descriptors if they have a `regex` property.

The changes to `breakLine_and_equalsAny.js` are superficial to appease eslint.