linearLex_oc_noErr and TokenizedInputElement
`linearLex_oc_noErr.js` is yet another version of linearLex (possibly the final one, or I might rename it to something else later) which changes the way invalid strings are handled when lexing. Instead of throwing errors when an issue is encountered, the lexer creates an error token representing the type of issue and then moves on, parsing the rest of the string.

This allows the output to be used for applications that don't necessarily require valid syntax, such as syntax highlight or auto-completion. This also allows for different issues to be handled more flexibly and gracefully, as each type of error token is assigned an identifying Symbol that can be used to customize behavior for different errors.

The output of linearLex is now an object containing the properties `valid`, `tokens`, and `errorTokens`. `valid` is a boolean representing whether the parameter string was a valid match for the given grammar. `tokens` includes the sequence of all tokens from the parameter string, including any error tokens produced along the way. Error tokens are included within the sequence in the location where the error was determined to have occurred. For example, an `UNMATCHED_CLOSER` token would appear just before the corresponding closing-parenthetical token, but an `UNMATCHED_OPENER` token will appear at the end of the list of tokens, since it's only known for sure that there are unclosed parentheticals once the whole string has been read. `errorTokens` only includes error tokens, as a convenience for applications which are only interested in information about errors.

`TokenizedInputElement.js` uses `linearLex_oc_noErr` to create a wrapper around input an input element which maintains a tokenization of its value, and which provides a few utility functions for interacting with the element based on its tokenization. Mainly, the `selectionTokens` method returns a list of tokens included in the current selection of the input element, even if some of those tokens straddle the boundary of the selection (and even if the selection is just a cursor, i.e. 0-width). The `replaceSelectedTokens` method takes the same selection and replaces it in the input element, which is particularly useful for auto-completion.