linearLex_oc_noErr bug-fix
Previously, there was a bug where, if a substring could not be matched by ANY type in the grammar, and the previous token's type specified a subset of all types in it's `next` property, the lexer would get caught in an infinite loop. I fixed this by adding a `nonMatchingTypes` array, which tracks which types have been tried for a match against the current substring., regardless of which ones were allowed.

This change was carried over to `TokenizedInputElement.js` as well.